% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{deck_stop_server}
\alias{deck_stop_server}
\title{Stop a running deckifyr web server}
\usage{
deck_stop_server(server)
}
\arguments{
\item{server}{A \code{deckifyr_server} object returned by \code{\link[=deck_serve]{deck_serve()}}.}
}
\value{
\code{server}, invisibly.
}
\description{
Kills the background process started by \code{\link[=deck_serve]{deck_serve()}}.
}
\examples{
\dontrun{
server <- deck_serve(project = "my-deck")
deck_stop_server(server)
}
}
