% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides.R
\name{deck_remove_slide}
\alias{deck_remove_slide}
\title{Remove a slide from a presentation.yaml}
\usage{
deck_remove_slide(presentation, id)
}
\arguments{
\item{presentation}{Path to \code{presentation.yaml}.}

\item{id}{The slide's id.}
}
\value{
The CLI result (invisibly).
}
\description{
Remove a slide from a presentation.yaml
}
